requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
mediapipe==0.10.14
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
scipy==1.12.0
scikit-image==0.22.0
cairosvg==2.7.1
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator==2.1.0
